requirements/base.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements/base.in
#
argon2-cffi==20.1.0
    # via -r requirements/base.in
beautifulsoup4==4.8.2
    # via wagtail
certifi==2020.12.5
    # via
    #   elasticsearch
    #   requests
cffi==1.14.5
    # via
    #   argon2-cffi
    #   cryptography
chardet==4.0.0
    # via requests
cryptography==3.4.7
    # via pyjwt
defusedxml==0.7.1
    # via python3-openid
django-allauth==0.44.0
    # via -r requirements/base.in
django-appconf==1.0.4
    # via django-compressor
django-cogwheels==0.3
    # via wagtailmenus
django-colorful==1.3
    # via puput
django-compressor==2.4
    # via -r requirements/base.in
django-controlled-vocabulary==0.11
    # via django-dublincore-resource
django-crispy-forms==1.11.1
    # via -r requirements/base.in
django-dublincore-resource==0.1.4
    # via -r requirements/base.in
django-el-pagination==3.3.0
    # via puput
django-elasticsearch-dsl==7.2.0
    # via -r requirements/base.in
django-environ==0.4.5
    # via -r requirements/base.in
django-kdl-wagtail==0.9.1
    # via -r requirements/base.in
django-model-utils==4.1.1
    # via -r requirements/base.in
django-modelcluster==5.1
    # via wagtail
django-ranged-response==0.2.0
    # via django-simple-captcha
django-simple-captcha==0.5.14
    # via -r requirements/base.in
django-social-share==2.2.1
    # via puput
django-taggit==1.4.0
    # via wagtail
django-treebeard==4.5.1
    # via wagtail
django==2.2.23
    # via
    #   -r requirements/base.in
    #   django-allauth
    #   django-appconf
    #   django-colorful
    #   django-controlled-vocabulary
    #   django-dublincore-resource
    #   django-el-pagination
    #   django-model-utils
    #   django-ranged-response
    #   django-simple-captcha
    #   django-taggit
    #   django-treebeard
    #   djangorestframework
    #   puput
    #   wagtail
djangorestframework==3.12.4
    # via wagtail
draftjs-exporter==2.1.7
    # via wagtail
elasticsearch-dsl==7.3.0
    # via
    #   -r requirements/base.in
    #   django-elasticsearch-dsl
elasticsearch==7.12.1
    # via elasticsearch-dsl
html5lib==1.1
    # via wagtail
idna==2.10
    # via requests
kdl-wagtail-page==0.1.2
    # via -r requirements/base.in
l18n==2020.6.1
    # via wagtail
oauthlib==3.1.0
    # via requests-oauthlib
pillow==7.2.0
    # via
    #   -r requirements/base.in
    #   django-simple-captcha
    #   wagtail
puput==1.1.1
    # via -r requirements/base.in
pycparser==2.20
    # via cffi
pyjwt[crypto]==2.1.0
    # via django-allauth
python-dateutil==2.8.1
    # via elasticsearch-dsl
python-slugify==4.0.1
    # via -r requirements/base.in
python3-openid==3.2.0
    # via django-allauth
pytz==2021.1
    # via
    #   -r requirements/base.in
    #   django
    #   django-modelcluster
    #   l18n
rcssmin==1.0.6
    # via
    #   -r requirements/base.in
    #   django-compressor
requests-oauthlib==1.3.0
    # via django-allauth
requests==2.25.1
    # via
    #   django-allauth
    #   requests-oauthlib
    #   wagtail
rjsmin==1.1.0
    # via django-compressor
six==1.16.0
    # via
    #   argon2-cffi
    #   django-compressor
    #   django-elasticsearch-dsl
    #   django-simple-captcha
    #   elasticsearch-dsl
    #   html5lib
    #   l18n
    #   python-dateutil
soupsieve==2.2.1
    # via beautifulsoup4
sqlparse==0.4.1
    # via django
text-unidecode==1.3
    # via python-slugify
unidecode==1.2.0
    # via wagtail
urllib3==1.26.4
    # via
    #   django-controlled-vocabulary
    #   elasticsearch
    #   requests
wagtail==2.9.3
    # via
    #   -r requirements/base.in
    #   django-kdl-wagtail
    #   kdl-wagtail-page
    #   puput
wagtailmenus==3.0.2
    # via -r requirements/base.in
webencodings==0.5.1
    # via html5lib
whitenoise==5.2.0
    # via -r requirements/base.in
willow==1.3
    # via wagtail
xlsxwriter==1.4.3
    # via wagtail
